Fix bug with prelude sqrt function incorrectly being labelled as Int -> Float instead of Float -> Float

+++ umber_std/prelude.ml
@@ -63,7 +63,7 @@ let names = {|
                    (Local
                     ((typ
                       (Scheme
-                       (Function ((Type_app Int ())) (Type_app Float ())))))))
+                       (Function ((Type_app Float ())) (Type_app Float ())))))))
                   (print
                    (Local
                     ((typ
